+++ src/main/java/com/example/newsfeed/controller/boardController/BoardController.java
@@ -6,7 +6,7 @@ import com.example.newsfeed.dto.boardDto.request.UpdateBoardRequestDto;
 import com.example.newsfeed.dto.boardDto.response.BoardResponseDto;
 import com.example.newsfeed.dto.boardDto.response.BoardsResponseDto;
 import com.example.newsfeed.dto.boardDto.response.UserBoardResponseDto;
-import com.example.newsfeed.service.boardService.BoardService;
+import com.example.newsfeed.service.board.BoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
