+++ src/main/java/com/example/newsfeed/entity/board/BoardLikes.java
@@ -1,6 +1,5 @@
-package com.example.newsfeed.entity.boardEntity;
+package com.example.newsfeed.entity.board;
 
-import com.example.newsfeed.common.entity.BaseEntity;
 import com.example.newsfeed.entity.userEntity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -8,15 +7,13 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Table(name = "comment")
+@Table(name = "board_likes")
 @NoArgsConstructor
-public class Comment extends BaseEntity {
+public class BoardLikes {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    private String contents;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "board_id")
@@ -26,4 +23,9 @@ public class Comment extends BaseEntity {
     @JoinColumn(name = "user_id")
     private User user;
 
+    public BoardLikes(Board board, User user){
+        this.board =board;
+        this.user = user;
+        this.board.like();
+    }
 }
