+++ src/main/java/com/example/newsfeed/entity/board/Comment.java
@@ -1,5 +1,6 @@
-package com.example.newsfeed.entity.boardEntity;
+package com.example.newsfeed.entity.board;
 
+import com.example.newsfeed.common.entity.BaseEntity;
 import com.example.newsfeed.entity.userEntity.User;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,24 +8,22 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Table(name = "comment_likes")
+@Table(name = "comment")
 @NoArgsConstructor
-public class CommentLikes {
+public class Comment extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    private String contents;
+
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
+    @JoinColumn(name = "board_id")
+    private Board board;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
-    public CommentLikes(Comment comment, User user){
-        this.comment = comment;
-        this.user = user;
-    }
 }
