+++ src/main/java/com/example/newsfeed/entity/board/CommentLikes.java
@@ -1,4 +1,4 @@
-package com.example.newsfeed.entity.boardEntity;
+package com.example.newsfeed.entity.board;
 
 import com.example.newsfeed.entity.userEntity.User;
 import jakarta.persistence.*;
@@ -7,25 +7,24 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@Table(name = "board_likes")
+@Table(name = "comment_likes")
 @NoArgsConstructor
-public class BoardLikes {
+public class CommentLikes {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "board_id")
-    private Board board;
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
     private User user;
 
-    public BoardLikes(Board board, User user){
-        this.board =board;
+    public CommentLikes(Comment comment, User user){
+        this.comment = comment;
         this.user = user;
-        this.board.like();
     }
 }
